requirements.txt
alembic==1.4.2
click==7.1.2
dominate==2.5.1
Flask==1.1.2
Flask-Bootstrap4==4.0.2
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.2
MarkupSafe==1.1.1
python-dateutil==2.8.1
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.3.17
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.1
